feature/16 refactor: 오타 수정
